% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_result}
\alias{check_result}
\title{Assistance function to check job status}
\usage{
check_result(url = "https://secure.holistics.io/", api_key, job_id)
}
\arguments{
\item{url}{URL to access your Holistics app. If nothing is provided, this will default to \url{https://secure.holistics.io/}}

\item{api_key}{API key to access Holistics API}

\item{job_id}{Report export's job ID, as returned by submit_request()}
}
\value{
None. If job succeeds, continue, If job fails, it will raise an error.
}
\description{
Assistance function to check job status
}
