% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_result}
\alias{download_result}
\title{Assistance function to download report result}
\usage{
download_result(url = "https://secure.holistics.io/", api_key, job_id)
}
\arguments{
\item{url}{URL to access your Holistics app. If nothing is provided, this will default to \url{https://secure.holistics.io/}}

\item{api_key}{API key to access Holistics API. Your Holistics user must be allowed to use API access first.}

\item{job_id}{Report export's job ID, as returned by submit_request()}
}
\value{
A dataframe containing report's data
}
\description{
Submit a GET request to download result created by the export job.
}
