% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{submit_request}
\alias{submit_request}
\title{Assistance function to submit export request}
\usage{
submit_request(url = "https://secure.holistics.io/", api_key, report_id,
  filters = list())
}
\arguments{
\item{url}{URL to access your Holistics app. If nothing is provided, this will default to \url{https://secure.holistics.io/}}

\item{api_key}{API key to access Holistics API. Your Holistics user must be allowed to use API access first.}

\item{report_id}{Can be the ID number, or the full string containing ID and the name of the report. \cr}

\item{filters}{A list of filters to apply to the requested report. These must be the filters that you already included in the report. \cr
\itemize{
  \item If the filter is of Dropdown type, fitler values must be the "key", not the display values that end-user sees in the report.
  \item Text or datetime values must be enclosed in quotation marks.
}
Please refer to example below on how translate this sample URL to a data export call: \cr
\code{https://secure.holistics.io/queries/1846-users-growth/queries?start_date=2019-01-01&end_date=2019-01-31&status=active}}
}
\value{
Report export's job ID
}
\description{
Submit a GET request to create data export
}
\examples{
\dontrun{
submit_request(url = "https://secure.holistics.io/"
                      , api_key = "api_key"
                      , report_id = "1846-users-growth"
                      , filters = list(
                          start_date = "2019-01-01"
                          , end_date = "2019-01-31"
                          , status = "active")
                       )
}

}
