% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolisticsR.R
\docType{package}
\name{Holistics-package}
\alias{Holistics}
\alias{Holistics-package}
\title{HolisticsR: R package to interact with Holistics API}
\description{
HolisticsR aims to help users integrate Holistics reports into their own data pipeline.
}
\details{
To learn more about Holistics' API, please visit: \url{https://docs.holistics.io/reference#holistics-api}
}
